Extract Finding Detail route surface
