+++ frontend/src/workbench/WorkbenchShell.tsx
@@ -2,7 +2,6 @@ import { Link, useLocation, useNavigate } from "@tanstack/react-router"
 import {
   Activity,
   AlertTriangle,
-  ArrowLeft,
   BarChart3,
   Database,
   Gauge,
@@ -20,8 +19,6 @@ import {
   ApiTokensService,
   type FindingDetailPublic,
   type FindingExplanationPublic,
-  type FindingOccurrencePublic,
-  type FindingPriority,
   type FindingPublic,
   FindingsService,
   type ImportParseErrorPublic,
@@ -48,10 +45,7 @@ import {
   type WorkbenchStatus,
 } from "../api-client"
 import { clearAccessToken, getAccessToken } from "../auth"
-import {
-  ProductAppShell,
-  type WorkbenchPath,
-} from "../components/app/AppShell"
+import { ProductAppShell, type WorkbenchPath } from "../components/app/AppShell"
 import {
   FindingsByPriorityChart,
   RiskTrendChart,
@@ -62,13 +56,9 @@ import {
   TopRemediationQueue,
 } from "../components/dashboard"
 import {
-  CvssBadge,
-  EpssBadge,
-  FindingStatusBadge,
-  KevBadge,
-  PriorityBadge,
-  RiskScore,
-} from "../components/risk"
+  demoFindingDetailForId,
+  demoFindingExplanationForDetail,
+} from "../components/finding-detail/finding-detail-model"
 import {
   type DataQualityNoticeItem,
   EmptyState,
@@ -77,23 +67,6 @@ import {
 } from "../components/states"
 import { Badge } from "../components/ui/badge"
 import { Button } from "../components/ui/button"
-import { Skeleton } from "../components/ui/skeleton"
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "../components/ui/table"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "../components/ui/tabs"
-import {
-  VpwSurface,
-  VpwSurfaceBody,
-  VpwSurfaceDescription,
-  VpwSurfaceHeader,
-  VpwSurfaceTitle,
-} from "../components/vpw"
 import {
   type ApiTokenScope,
   apiTokenScopeOptions,
@@ -119,8 +92,6 @@ import {
   analysisRunIdFromError,
   apiErrorDetail,
   apiErrorMessage,
-  arrayRecords,
-  joinedValues,
   objectRecord,
   parseErrorsFromError,
   stringValue,
@@ -134,27 +105,14 @@ import {
   topServicesByRiskChartData,
 } from "../lib/chart-data"
 import {
-  DEMO_FINDING_ATTACK_CONTEXTS,
-  DEMO_FINDINGS,
-  DEMO_PROJECT,
-} from "../lib/demo-data"
-import {
   formatCacheAge,
   formatProviderFreshness,
   providerDataQualityNotes,
   providerSnapshotSummary,
 } from "../lib/provider-format"
+import { runStatusLabel } from "../lib/risk-format"
+import { formatLabel as labelize } from "../lib/ui-copy"
 import {
-  formatEpss,
-  formatKev,
-  formatNullableNumber,
-  runStatusLabel,
-} from "../lib/risk-format"
-import { formatLabel as labelize, optionalText } from "../lib/ui-copy"
-import { cn } from "../lib/utils"
-import {
-  type FindingWaiverEvidence,
-  findingWaiverEvidence,
   validateWaiverForm,
   type WaiverFormState,
   waiverFormDefaults,
@@ -170,6 +128,11 @@ const RiskOperationsDashboard = lazy(() =>
 const RemediationQueue = lazy(() =>
   import("../components/findings/RemediationQueue").then((module) => ({
     default: module.RemediationQueue,
+  })),
+)
+const FindingDetailRoute = lazy(() =>
+  import("../components/finding-detail/FindingDetailRoute").then((module) => ({
+    default: module.FindingDetailRoute,
   })),
 )
 const ImportsWorkbench = lazy(() =>
@@ -202,20 +165,6 @@ const WaiversWorkbench = lazy(() =>
     default: module.WaiversWorkbench,
   })),
 )
-
-type FindingAttackContext = NonNullable<FindingDetailPublic["attack_context"]>
-
-type FindingDecisionReason = {
-  detail: string
-  label: string
-  tone: "critical" | "warning" | "info" | "positive"
-}
-
-type FindingDetailRow = {
-  detail?: string
-  label: string
-  value: string
-}
 
 function _priorityCount(
   summary: ProjectDecisionSummaryPublic | null,
@@ -474,712 +423,6 @@ function _failedRunCause(
   )
 }
 
-const decisionReasonCopy: Record<string, { label: string; detail: string }> = {
-  "asset.context": {
-    detail: "Asset context influences operational priority.",
-    label: "Asset context",
-  },
-  "asset.context_unknown": {
-    detail:
-      "Asset context is missing and must be validated before final scheduling.",
-    label: "Asset context unknown",
-  },
-  "operational.score": {
-    detail: "Combined signals determine the operational remediation score.",
-    label: "Operational score",
-  },
-  "priority.critical.epss_cvss": {
-    detail: "EPSS and CVSS together indicate critical remediation urgency.",
-    label: "Critical EPSS and CVSS",
-  },
-  "priority.high.cvss": {
-    detail: "CVSS indicates high impact severity.",
-    label: "High CVSS",
-  },
-  "priority.high.epss": {
-    detail: "EPSS indicates elevated exploitation probability.",
-    label: "High EPSS",
-  },
-  "priority.kev.known_exploited": {
-    detail: "Known exploited vulnerability signal is present.",
-    label: "Known exploited vulnerability",
-  },
-  "priority.medium.cvss": {
-    detail: "CVSS contributes meaningful impact severity.",
-    label: "Medium CVSS signal",
-  },
-  "priority.medium.epss": {
-    detail: "EPSS contributes exploitation probability context.",
-    label: "Medium EPSS signal",
-  },
-}
-
-function humanizeDecisionReasonText(value: string | null | undefined) {
-  if (!value) {
-    return value
-  }
-  return Object.entries(decisionReasonCopy).reduce(
-    (text, [code, copy]) =>
-      text.replaceAll(code, copy.detail.replace(/\.$/, "")),
-    value,
-  )
-}
-
-function decisionReasonLabel(value: string | null | undefined) {
-  if (!value) {
-    return "Reason"
-  }
-  return decisionReasonCopy[value]?.label ?? labelize(value)
-}
-
-function decisionReasonDetail(
-  code: string | null | undefined,
-  detail: string | null | undefined,
-) {
-  return (
-    humanizeDecisionReasonText(detail) ?? decisionReasonCopy[code ?? ""]?.detail
-  )
-}
-
-function _findingOverviewCards(finding: FindingDetailPublic) {
-  return [
-    {
-      detail:
-        finding.epss === null || finding.epss === undefined
-          ? "EPSS provider data missing"
-          : "FIRST EPSS probability",
-      label: "EPSS",
-      value: formatEpss(finding.epss),
-    },
-    {
-      detail:
-        finding.cvss_base_score === null ||
-        finding.cvss_base_score === undefined
-          ? "CVSS provider data missing"
-          : "NVD base score",
-      label: "CVSS",
-      value: formatNullableNumber(finding.cvss_base_score),
-    },
-    {
-      detail: finding.in_kev ? "CISA KEV matched" : "No KEV match recorded",
-      label: "KEV",
-      value: formatKev(finding.in_kev),
-    },
-    {
-      detail: joinedValues([
-        finding.owner,
-        finding.business_service,
-        labelize(finding.asset_environment),
-        labelize(finding.asset_criticality),
-        labelize(finding.exposure),
-        finding.asset_target_ref,
-      ]),
-      label: "Asset",
-      value: findingAssetLabel(finding),
-    },
-  ]
-}
-
-function findingOccurrenceRows(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-): Array<Partial<FindingOccurrencePublic> & Record<string, unknown>> {
-  if (finding?.occurrences?.length) {
-    return finding.occurrences
-  }
-  const explanationPayload = objectRecord(explanation?.explanation)
-  const explanationProvenance = objectRecord(explanationPayload.provenance)
-  const findingProvenance = objectRecord(
-    objectRecord(finding?.explanation_json).provenance,
-  )
-  return arrayRecords(
-    explanationProvenance.occurrences ?? findingProvenance.occurrences,
-  ).map((occurrence, index) => ({
-    ...occurrence,
-    id: stringValue(occurrence.id) ?? `occurrence-${index + 1}`,
-  }))
-}
-
-function findingWhyText(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-) {
-  const decisionExplanation = objectRecord(explanation?.decision_explanation)
-  const fallback = "No priority explanation has been recorded for this finding."
-  return (
-    humanizeDecisionReasonText(
-      stringValue(decisionExplanation.human_readable) ??
-        stringValue(decisionExplanation.summary) ??
-        explanation?.rationale ??
-        finding?.rationale ??
-        fallback,
-    ) ?? fallback
-  )
-}
-
-function findingRecommendedAction(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-) {
-  const decisionGuidance = objectRecord(explanation?.decision_guidance)
-  return (
-    stringValue(decisionGuidance.recommended_action) ??
-    explanation?.recommended_action ??
-    finding?.recommended_action ??
-    "No recommended action has been recorded."
-  )
-}
-
-function findingComponentDetailLabel(finding: FindingDetailPublic | null) {
-  if (!finding) {
-    return "N.A."
-  }
-  const name = optionalText(finding.component_name)
-  return finding.component_version
-    ? `${name} ${finding.component_version}`
-    : name
-}
-
-function findingAssetServiceDetailLabel(finding: FindingDetailPublic | null) {
-  if (!finding) {
-    return "N.A."
-  }
-  return joinedValues([
-    finding.business_service,
-    finding.asset_name,
-    finding.asset_key,
-    finding.asset_target_ref,
-  ])
-}
-
-function findingOwnerDetailLabel(
-  finding: FindingDetailPublic | null,
-  occurrences: Array<
-    Partial<FindingOccurrencePublic> & Record<string, unknown>
-  >,
-) {
-  return (
-    finding?.owner ?? stringValue(occurrences[0]?.asset_owner) ?? "Unassigned"
-  )
-}
-
-function findingSlaLabel(priority: FindingPriority | undefined) {
-  switch (priority) {
-    case "critical":
-      return "24 hours"
-    case "high":
-      return "7 days"
-    case "medium":
-      return "30 days"
-    case "low":
-      return "90 days"
-    default:
-      return "Define during triage"
-  }
-}
-
-function findingNextStepLabel(finding: FindingDetailPublic | null) {
-  switch (finding?.status) {
-    case "open":
-      return "Assign remediation work and start fix validation."
-    case "in_review":
-      return "Complete technical review and confirm the remediation path."
-    case "remediating":
-      return "Verify the fix, then update evidence and status."
-    case "fixed":
-      return "Confirm scanner closure and keep evidence for reporting."
-    case "accepted":
-      return "Track accepted risk until the next review date."
-    case "suppressed":
-      return "Verify the VEX or suppression scope still applies."
-    default:
-      return "Confirm ownership and record the next remediation step."
-  }
-}
-
-function isInternetFacingExposure(value: string | null | undefined) {
-  return value ? value.toLowerCase().includes("internet") : false
-}
-
-function isProductionEnvironment(value: string | null | undefined) {
-  return value ? /\bprod(uction)?\b/i.test(value.replaceAll("_", " ")) : false
-}
-
-function findingDecisionReasonRows(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-  attackContext: FindingAttackContext | null,
-) {
-  const rows: FindingDecisionReason[] = []
-
-  if (finding?.in_kev) {
-    rows.push({
-      detail:
-        "CISA KEV is recorded for this CVE, so exploitation is a confirmed prioritization signal.",
-      label: "CISA KEV listed",
-      tone: "critical",
-    })
-  }
-
-  if (finding?.epss !== null && finding?.epss !== undefined) {
-    rows.push({
-      detail:
-        finding.epss >= 0.7
-          ? `${formatEpss(finding.epss)} EPSS indicates elevated exploitation probability.`
-          : `${formatEpss(finding.epss)} EPSS is recorded for the decision model.`,
-      label: finding.epss >= 0.7 ? "High EPSS" : "EPSS recorded",
-      tone: finding.epss >= 0.7 ? "warning" : "info",
-    })
-  }
-
-  if (
-    finding?.cvss_base_score !== null &&
-    finding?.cvss_base_score !== undefined
-  ) {
-    rows.push({
-      detail:
-        finding.cvss_base_score >= 9
-          ? `CVSS ${formatNullableNumber(finding.cvss_base_score)} is critical impact severity.`
-          : `CVSS ${formatNullableNumber(finding.cvss_base_score)} contributes to the risk score.`,
-      label: finding.cvss_base_score >= 9 ? "Critical CVSS" : "CVSS recorded",
-      tone: finding.cvss_base_score >= 9 ? "critical" : "info",
-    })
-  }
-
-  if (isInternetFacingExposure(finding?.exposure)) {
-    rows.push({
-      detail: `${finding ? findingAssetLabel(finding) : "Asset"} is marked ${labelize(finding?.exposure)}, increasing remediation urgency.`,
-      label: "Internet-facing asset",
-      tone: "critical",
-    })
-  }
-
-  if (isProductionEnvironment(finding?.asset_environment)) {
-    rows.push({
-      detail: `Environment is marked ${labelize(finding?.asset_environment)}, so operational exposure is higher.`,
-      label: "Production service",
-      tone: "warning",
-    })
-  }
-
-  if (!attackContextEmptyState(attackContext)) {
-    rows.push({
-      detail: `${attackTechniqueRows(attackContext).length} reviewed ATT&CK technique mapping(s) are available for defensive context.`,
-      label: "ATT&CK / TTP context",
-      tone: "warning",
-    })
-  }
-
-  const defensiveNote = stringValue(attackContext?.defensive_note)
-  if (defensiveNote && /gap|missing|coverage|detect/i.test(defensiveNote)) {
-    rows.push({
-      detail: defensiveNote,
-      label: "Detection coverage gap",
-      tone: "warning",
-    })
-  }
-
-  const recommendedAction = findingRecommendedAction(finding, explanation)
-  if (recommendedAction !== "No recommended action has been recorded.") {
-    rows.push({
-      detail: recommendedAction,
-      label: "Fix or mitigation",
-      tone: "positive",
-    })
-  }
-
-  if (rows.length === 0) {
-    rows.push({
-      detail: findingWhyText(finding, explanation),
-      label: "Decision rationale",
-      tone: "info",
-    })
-  }
-
-  return rows
-}
-
-function findingReasonRows(explanation: FindingExplanationPublic | null) {
-  const decisionExplanation = objectRecord(explanation?.decision_explanation)
-  const reasons = arrayRecords(decisionExplanation.reasons)
-  return reasons.map((reason, index) => {
-    const code =
-      stringValue(reason.code) ??
-      stringValue(reason.signal) ??
-      stringValue(reason.source)
-    const detail =
-      decisionReasonDetail(
-        code,
-        stringValue(reason.message) ??
-          stringValue(reason.description) ??
-          stringValue(reason.detail) ??
-          stringValue(reason.value) ??
-          null,
-      ) ?? "Matched decision signal"
-    return {
-      detail,
-      label: code ? decisionReasonLabel(code) : `Reason ${index + 1}`,
-    }
-  })
-}
-
-function findingDataQualityRows(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-) {
-  const flags =
-    explanation?.data_quality_flags ??
-    arrayRecords(objectRecord(finding?.data_quality_json).flags)
-  return flags.map((flag, index) => ({
-    code: stringValue(flag.code) ?? "data_quality_flag",
-    key: [flag.source, flag.code, flag.message, index].join(":"),
-    message: stringValue(flag.message) ?? "Data quality flag recorded.",
-    severity: stringValue(flag.severity) ?? "info",
-    source: stringValue(flag.source) ?? "provider",
-  }))
-}
-
-function findingProviderGaps(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-) {
-  if (!finding) {
-    return []
-  }
-  const providerEvidence = objectRecord(
-    explanation?.provider_evidence ??
-      objectRecord(finding.explanation_json).provider_evidence,
-  )
-  const gaps: string[] = []
-  if (finding.epss === null || finding.epss === undefined) {
-    gaps.push("EPSS missing")
-  }
-  if (
-    finding.cvss_base_score === null ||
-    finding.cvss_base_score === undefined
-  ) {
-    gaps.push("CVSS missing")
-  }
-  if (Object.keys(providerEvidence).length === 0) {
-    gaps.push("Provider evidence missing")
-  }
-  return gaps
-}
-
-function findingEvidenceRows(
-  finding: FindingDetailPublic | null,
-  explanation: FindingExplanationPublic | null,
-  occurrences: Array<
-    Partial<FindingOccurrencePublic> & Record<string, unknown>
-  >,
-  dataQualityRows: ReturnType<typeof findingDataQualityRows>,
-  providerGaps: string[],
-): FindingDetailRow[] {
-  const firstOccurrence = occurrences[0]
-  const providerEvidence = objectRecord(
-    explanation?.provider_evidence ??
-      objectRecord(finding?.explanation_json).provider_evidence,
-  )
-  const evidence = objectRecord(finding?.evidence_json)
-  const providerSignalLabels = [
-    finding?.epss !== null && finding?.epss !== undefined ? "EPSS" : null,
-    finding?.cvss_base_score !== null && finding?.cvss_base_score !== undefined
-      ? "CVSS"
-      : null,
-    finding?.in_kev ? "CISA KEV" : null,
-  ].filter((label): label is string => Boolean(label))
-  const providerKeys = Object.keys(providerEvidence)
-  const artifactRef =
-    stringValue(evidence.report_artifact) ??
-    stringValue(evidence.report_reference) ??
-    stringValue(evidence.artifact_reference) ??
-    stringValue(firstOccurrence?.vex_source_path)
-
-  return [
-    {
-      detail:
-        providerGaps.length > 0
-          ? `Gaps: ${providerGaps.join(", ")}`
-          : "No provider gaps recorded for the stored finding signals.",
-      label: "Provider snapshot",
-      value:
-        providerKeys.length > 0
-          ? `${providerKeys.length} provider evidence field(s) recorded`
-          : providerSignalLabels.length > 0
-            ? providerSignalLabels.join(", ")
-            : "No provider snapshot recorded",
-    },
-    {
-      detail: optionalText(
-        stringValue(firstOccurrence?.source_record_id) ??
-          stringValue(firstOccurrence?.raw_reference) ??
-          stringValue(firstOccurrence?.source_id),
-      ),
-      label: "Input source",
-      value: optionalText(
-        stringValue(firstOccurrence?.source_format) ??
-          stringValue(firstOccurrence?.source) ??
-          stringValue(evidence.input_source),
-      ),
-    },
-    {
-      detail: optionalText(stringValue(firstOccurrence?.raw_severity)),
-      label: "Scanner evidence",
-      value: optionalText(
-        stringValue(firstOccurrence?.scanner) ??
-          stringValue(evidence.scanner) ??
-          stringValue(evidence.tool),
-      ),
-    },
-    {
-      detail:
-        dataQualityRows.length > 0
-          ? dataQualityRows.map((row) => labelize(row.code)).join(", ")
-          : "No data quality flags recorded.",
-      label: "Data quality notes",
-      value:
-        explanation?.data_quality_confidence ??
-        stringValue(objectRecord(finding?.data_quality_json).confidence) ??
-        "Recorded",
-    },
-    {
-      detail:
-        "Report or evidence bundle reference when supplied by input data.",
-      label: "Report / artifact references",
-      value: optionalText(artifactRef),
-    },
-  ]
-}
-
-function findingHistoryRows(
-  finding: FindingDetailPublic | null,
-  occurrences: Array<
-    Partial<FindingOccurrencePublic> & Record<string, unknown>
-  >,
-  waiverEvidence: FindingWaiverEvidence | null,
-): FindingDetailRow[] {
-  const vexStatus =
-    occurrences
-      .map((occurrence) => stringValue(occurrence.vex_status))
-      .find(Boolean) ?? null
-  const vexDetail =
-    occurrences
-      .map((occurrence) =>
-        joinedValues([
-          stringValue(occurrence.vex_justification),
-          stringValue(occurrence.vex_action_statement),
-          stringValue(occurrence.vex_match_type),
-        ]),
-      )
-      .find((value) => value !== "N.A.") ?? null
-
-  return [
-    {
-      detail: "Initial source occurrence recorded by Workbench.",
-      label: "First seen",
-      value: finding?.first_seen_at
-        ? formatDateTime(finding.first_seen_at)
-        : "N.A.",
-    },
-    {
-      detail: "Most recent source occurrence recorded by Workbench.",
-      label: "Last seen",
-      value: finding?.last_seen_at
-        ? formatDateTime(finding.last_seen_at)
-        : "N.A.",
-    },
-    {
-      detail: finding?.updated_at
-        ? `Last updated ${formatDateTime(finding.updated_at)}`
-        : "No status update timestamp recorded.",
-      label: "Status changes",
-      value: labelize(finding?.status),
-    },
-    {
-      detail:
-        waiverEvidence?.reason ??
-        vexDetail ??
-        "No waiver or VEX state recorded for this finding.",
-      label: "Waiver / VEX state",
-      value: finding?.waived
-        ? `Waived${waiverEvidence?.status ? ` (${labelize(waiverEvidence.status)})` : ""}`
-        : finding?.suppressed_by_vex
-          ? "Suppressed by VEX"
-          : vexStatus
-            ? labelize(vexStatus)
-            : "Not accepted",
-    },
-  ]
-}
-
-function demoFindingDetailForId(findingId: string) {
-  const demoFinding = DEMO_FINDINGS.find((finding) => finding.id === findingId)
-  if (!demoFinding) {
-    return null
-  }
-  const createdAt = demoFinding.created_at ?? "2025-04-01T00:00:00Z"
-  const lastSeenAt = demoFinding.last_seen_at ?? createdAt
-  return {
-    ...demoFinding,
-    asset_id: demoFinding.asset_id ?? null,
-    component_id: demoFinding.component_id ?? null,
-    created_at: createdAt,
-    data_quality_json: {
-      confidence: "Demo Preview",
-      flags: [
-        {
-          code: "demo_preview",
-          message:
-            "Sample finding detail used only when no real project data is available.",
-          severity: "info",
-          source: "demo",
-        },
-      ],
-    },
-    evidence_json: {
-      input_source: "Demo Preview sample queue",
-      scanner: "Demo import",
-    },
-    first_seen_at: demoFinding.first_seen_at ?? createdAt,
-    last_seen_at: lastSeenAt,
-    attack_context: DEMO_FINDING_ATTACK_CONTEXTS[demoFinding.id] ?? null,
-    occurrences: [
-      {
-        analysis_run_id: "demo-run-0001",
-        asset_business_service: demoFinding.business_service,
-        asset_exposure: demoFinding.exposure,
-        asset_owner: demoFinding.owner,
-        asset_ref:
-          demoFinding.asset_name ??
-          demoFinding.asset_key ??
-          demoFinding.business_service,
-        component_name: demoFinding.component_name,
-        component_version: demoFinding.component_version,
-        created_at: createdAt,
-        fix_version: stringValue(demoFinding.recommended_action),
-        id: `${demoFinding.id}-occurrence-1`,
-        purl: demoFinding.component_purl,
-        raw_severity: labelize(demoFinding.priority),
-        scanner: "Demo import",
-        source: "Demo Preview",
-        source_format: "Sample finding",
-        source_record_id: demoFinding.cve_id,
-        target_kind: "service",
-        target_ref: demoFinding.business_service,
-      },
-    ],
-    project_id: DEMO_PROJECT.id,
-    updated_at: demoFinding.updated_at ?? lastSeenAt,
-    vulnerability_id: demoFinding.vulnerability_id ?? demoFinding.cve_id,
-  } as FindingDetailPublic
-}
-
-function demoFindingExplanationForDetail(
-  finding: FindingDetailPublic,
-): FindingExplanationPublic {
-  return {
-    cve_id: finding.cve_id,
-    data_quality_confidence: "Demo Preview",
-    decision_explanation: {
-      human_readable: findingWhyText(finding, null),
-    },
-    decision_guidance: {
-      recommended_action: findingRecommendedAction(finding, null),
-    },
-    finding_id: finding.id,
-    priority: finding.priority ?? "medium",
-    priority_rank: finding.priority_rank ?? 0,
-    project_id: finding.project_id,
-    provider_evidence: {
-      demo_preview: true,
-      epss: finding.epss,
-      cvss_base_score: finding.cvss_base_score,
-      in_kev: finding.in_kev,
-    },
-    rationale: finding.rationale,
-    recommended_action: finding.recommended_action,
-    risk_score: finding.risk_score,
-  }
-}
-
-function attackTechniqueRows(context: FindingAttackContext | null) {
-  if (!context) {
-    return []
-  }
-  const techniques = context.techniques ?? []
-  if (techniques.length > 0) {
-    return techniques
-  }
-  return (context.mappings ?? []).map((mapping) => ({
-    confidence: mapping.confidence,
-    defensive_note: mapping.defensive_note,
-    name: mapping.technique_name,
-    rationale: mapping.rationale,
-    review_status: mapping.review_status,
-    source: mapping.source,
-    tactics: mapping.tactics ?? [],
-    technique_id: mapping.technique_id,
-    url: null,
-  }))
-}
-
-function attackTacticsLabel(values: string[] | null | undefined) {
-  return values && values.length > 0 ? values.join(", ") : "N.A."
-}
-
-function attackConfidenceLabel(value: string | null | undefined) {
-  return value ? labelize(value) : "Unknown"
-}
-
-function attackSourceLabel(
-  source: string | null | undefined,
-  context: FindingAttackContext | null,
-) {
-  const references = (context?.mappings ?? []).flatMap(
-    (mapping) => mapping.references ?? [],
-  )
-  if (
-    references.some((reference) =>
-      reference.toLowerCase().includes("local curated demo mapping"),
-    )
-  ) {
-    return "Local curated demo mapping"
-  }
-  if (source === "local-curated") {
-    return "Local curated mapping"
-  }
-  return source ?? null
-}
-
-function attackCoverageStatusLabel(context: FindingAttackContext | null) {
-  const note = (stringValue(context?.defensive_note) ?? "").toLowerCase()
-  if (!note) {
-    return "Not recorded"
-  }
-  if (/partial|unknown/.test(note)) {
-    return "Partial / unknown"
-  }
-  if (/gap|missing|validate|coverage|detect/.test(note)) {
-    return "Needs validation"
-  }
-  return "Recorded"
-}
-
-function defensiveActionItems(value: string | null | undefined) {
-  return (value ?? "")
-    .split(/\n+/)
-    .map((item) => item.trim())
-    .filter(Boolean)
-}
-
-function attackContextEmptyState(context: FindingAttackContext | null) {
-  return (
-    !context ||
-    context.mapped !== true ||
-    attackTechniqueRows(context).length === 0
-  )
-}
-
 function numericFilterValue(value: string) {
   const trimmed = value.trim()
   if (!trimmed) {
@@ -1191,22 +434,6 @@ function numericFilterValue(value: string) {
 
 function hasActiveFindingFilters(filters: FindingFilters) {
   return Object.values(filters).some((value) => value.trim() !== "")
-}
-
-function _findingComponentLabel(finding: FindingPublic) {
-  const name = optionalText(finding.component_name)
-  return finding.component_version
-    ? `${name} ${finding.component_version}`
-    : name
-}
-
-function findingAssetLabel(finding: FindingPublic) {
-  return (
-    finding.asset_name ??
-    finding.asset_key ??
-    finding.business_service ??
-    "N.A."
-  )
 }
 
 function _metadataRows(value: unknown) {
@@ -1633,65 +860,6 @@ export function WorkbenchShell({
   )
   const activeFindingFilters =
     hasActiveFindingFilters(findingFilters) || Boolean(findingAssetId)
-  const detailOccurrences = findingOccurrenceRows(
-    findingDetail,
-    findingExplanation,
-  )
-  const detailDataQualityRows = findingDataQualityRows(
-    findingDetail,
-    findingExplanation,
-  )
-  const detailProviderGaps = findingProviderGaps(
-    findingDetail,
-    findingExplanation,
-  )
-  const detailReasonRows = findingReasonRows(findingExplanation)
-  const detailAttackContext = findingDetail?.attack_context ?? null
-  const detailAttackTechniques = attackTechniqueRows(detailAttackContext)
-  const detailAttackPrimaryTechnique = detailAttackTechniques[0] ?? null
-  const detailAttackActionItems = defensiveActionItems(
-    detailAttackPrimaryTechnique?.defensive_note,
-  )
-  const detailAttackCoverageStatus =
-    attackCoverageStatusLabel(detailAttackContext)
-  const detailAttackTacticLabel = attackTacticsLabel(
-    detailAttackPrimaryTechnique?.tactics ?? detailAttackContext?.tactics,
-  )
-  const detailAttackTechniqueId =
-    detailAttackPrimaryTechnique?.technique_id ?? "Technique"
-  const detailAttackTechniqueName =
-    detailAttackPrimaryTechnique?.name ?? "Technique"
-  const detailAttackTechniqueLabel = `${detailAttackTechniqueId} ${detailAttackTechniqueName}`
-  const detailAttackSource = attackSourceLabel(
-    detailAttackPrimaryTechnique?.source ?? detailAttackContext?.source,
-    detailAttackContext,
-  )
-  const detailAttackRationale =
-    detailAttackPrimaryTechnique?.rationale ?? detailAttackContext?.rationale
-  const detailAttackEmpty = attackContextEmptyState(detailAttackContext)
-  const detailWaiverEvidence = findingWaiverEvidence(findingDetail)
-  const detailDecisionReasons = findingDecisionReasonRows(
-    findingDetail,
-    findingExplanation,
-    detailAttackContext,
-  )
-  const detailEvidenceRows = findingEvidenceRows(
-    findingDetail,
-    findingExplanation,
-    detailOccurrences,
-    detailDataQualityRows,
-    detailProviderGaps,
-  )
-  const detailHistoryRows = findingHistoryRows(
-    findingDetail,
-    detailOccurrences,
-    detailWaiverEvidence,
-  )
-  const isDemoFindingDetail = Boolean(
-    findingDetail &&
-      (findingDetail.project_id === DEMO_PROJECT.id ||
-        findingDetail.id.startsWith("demo-")),
-  )
   const selectedReportRun =
     projectRuns.find((run) => run.id === selectedRunId) ?? null
   const reportActionsEnabled =
@@ -3321,1046 +2489,16 @@ export function WorkbenchShell({
                   waiversLoading={waiversLoading || governanceLoading}
                 />
               ) : isFindingDetail ? (
-                <section
-                  className="finding-decision-workflow"
-                  aria-label="Finding priority decision"
-                >
-                  <div className="finding-detail-backbar">
-                    <Button variant="outline" size="sm" asChild>
-                      <Link to="/findings">
-                        <ArrowLeft aria-hidden="true" size={16} />
-                        <span>Back to Findings</span>
-                      </Link>
-                    </Button>
-                  </div>
-
-                  {findingDetailError ? (
-                    <p
-                      className="text-sm text-destructive rounded-md border border-destructive/20 bg-destructive/10 px-3 py-2"
-                      role="alert"
-                    >
-                      {findingDetailError}
-                    </p>
-                  ) : null}
-                  {findingExplanationWarning ? (
-                    <p
-                      className="text-sm text-destructive rounded-md border border-destructive/20 bg-destructive/10 px-3 py-2"
-                      role="alert"
-                    >
-                      {findingExplanationWarning}
-                    </p>
-                  ) : null}
-                  {findingDetailLoading ? (
-                    <section
-                      aria-label="Loading finding detail"
-                      className="finding-decision-loading"
-                      role="status"
-                    >
-                      <Skeleton className="h-8 w-40" />
-                      <Skeleton className="h-36 w-full" />
-                      <div className="grid grid-cols-1 gap-3 md:grid-cols-4">
-                        <Skeleton className="h-24 w-full" />
-                        <Skeleton className="h-24 w-full" />
-                        <Skeleton className="h-24 w-full" />
-                        <Skeleton className="h-24 w-full" />
-                      </div>
-                    </section>
-                  ) : null}
-
-                  {!findingDetailLoading &&
-                  !findingDetailError &&
-                  findingDetail ? (
-                    <>
-                      {isDemoFindingDetail ? (
-                        <section
-                          aria-label="Demo Preview"
-                          className="finding-demo-preview"
-                        >
-                          <Badge className="border-amber-200 bg-amber-50 text-amber-800">
-                            Demo Preview
-                          </Badge>
-                          <span>
-                            Sample evidence for the decision workflow. Not real
-                            production data.
-                          </span>
-                        </section>
-                      ) : null}
-
-                      <section
-                        className="finding-decision-hero"
-                        aria-label="Finding decision hero"
-                      >
-                        <div className="finding-decision-hero-copy">
-                          <Badge className="finding-decision-kicker">
-                            Why this priority?
-                          </Badge>
-                          <h2>{findingDetail.cve_id}</h2>
-                          <p>
-                            {findingWhyText(findingDetail, findingExplanation)}
-                          </p>
-                          <div className="finding-decision-badges">
-                            <PriorityBadge priority={findingDetail.priority} />
-                            <FindingStatusBadge status={findingDetail.status} />
-                            <KevBadge matched={findingDetail.in_kev} />
-                          </div>
-                        </div>
-                        <ul
-                          className="finding-decision-hero-metrics"
-                          aria-label="Risk indicators"
-                        >
-                          <li className="finding-hero-metric-primary">
-                            <span>Risk Score</span>
-                            <strong>
-                              {formatNullableNumber(findingDetail.risk_score)}
-                            </strong>
-                            <small>Operational remediation priority</small>
-                          </li>
-                          <li>
-                            <span>EPSS</span>
-                            <EpssBadge value={findingDetail.epss} />
-                            <small>Exploit probability signal</small>
-                          </li>
-                          <li>
-                            <span>CVSS</span>
-                            <CvssBadge value={findingDetail.cvss_base_score} />
-                            <small>Impact severity signal</small>
-                          </li>
-                          <li className="finding-hero-metric-decision">
-                            <span>Decision</span>
-                            <RiskScore value={findingDetail.risk_score} />
-                            <small>
-                              {findingSlaLabel(findingDetail.priority)} SLA
-                            </small>
-                          </li>
-                        </ul>
-                      </section>
-
-                      <section
-                        className="finding-decision-fact-grid"
-                        aria-label="Finding scope"
-                      >
-                        <VpwSurface>
-                          <VpwSurfaceHeader>
-                            <VpwSurfaceDescription>
-                              Component
-                            </VpwSurfaceDescription>
-                            <VpwSurfaceTitle
-                              title={findingComponentDetailLabel(findingDetail)}
-                            >
-                              {findingComponentDetailLabel(findingDetail)}
-                            </VpwSurfaceTitle>
-                          </VpwSurfaceHeader>
-                          <VpwSurfaceBody>
-                            <p
-                              title={optionalText(findingDetail.component_purl)}
-                            >
-                              {optionalText(findingDetail.component_purl)}
-                            </p>
-                          </VpwSurfaceBody>
-                        </VpwSurface>
-                        <VpwSurface>
-                          <VpwSurfaceHeader>
-                            <VpwSurfaceDescription>
-                              Asset / Service
-                            </VpwSurfaceDescription>
-                            <VpwSurfaceTitle
-                              title={findingAssetServiceDetailLabel(
-                                findingDetail,
-                              )}
-                            >
-                              {findingAssetServiceDetailLabel(findingDetail)}
-                            </VpwSurfaceTitle>
-                          </VpwSurfaceHeader>
-                          <VpwSurfaceBody>
-                            <p>{labelize(findingDetail.exposure)}</p>
-                          </VpwSurfaceBody>
-                        </VpwSurface>
-                        <VpwSurface>
-                          <VpwSurfaceHeader>
-                            <VpwSurfaceDescription>Owner</VpwSurfaceDescription>
-                            <VpwSurfaceTitle>
-                              {findingOwnerDetailLabel(
-                                findingDetail,
-                                detailOccurrences,
-                              )}
-                            </VpwSurfaceTitle>
-                          </VpwSurfaceHeader>
-                          <VpwSurfaceBody>
-                            <p>{labelize(findingDetail.asset_environment)}</p>
-                          </VpwSurfaceBody>
-                        </VpwSurface>
-                        <VpwSurface className="finding-recommendation-card">
-                          <VpwSurfaceHeader>
-                            <VpwSurfaceDescription>
-                              Primary recommendation
-                            </VpwSurfaceDescription>
-                            <VpwSurfaceTitle>
-                              {findingRecommendedAction(
-                                findingDetail,
-                                findingExplanation,
-                              )}
-                            </VpwSurfaceTitle>
-                          </VpwSurfaceHeader>
-                        </VpwSurface>
-                      </section>
-
-                      <section
-                        className="finding-decision-main-grid"
-                        aria-label="Risk to decision"
-                      >
-                        <VpwSurface className="finding-decision-card finding-analysis-card">
-                          <VpwSurfaceHeader>
-                            <div className="finding-card-heading">
-                              <div>
-                                <VpwSurfaceDescription>
-                                  Risk to Decision
-                                </VpwSurfaceDescription>
-                                <VpwSurfaceTitle>
-                                  Why this priority?
-                                </VpwSurfaceTitle>
-                              </div>
-                              <Badge variant="outline">
-                                Score{" "}
-                                {formatNullableNumber(findingDetail.risk_score)}
-                              </Badge>
-                            </div>
-                          </VpwSurfaceHeader>
-                          <VpwSurfaceBody>
-                            <p className="finding-decision-lead">
-                              {findingWhyText(
-                                findingDetail,
-                                findingExplanation,
-                              )}
-                            </p>
-                            <ol
-                              aria-label="Risk to decision chain"
-                              className="finding-decision-chain"
-                            >
-                              <li>Finding</li>
-                              <li>Priority</li>
-                              <li>Evidence</li>
-                              <li>Decision</li>
-                            </ol>
-                            <dl className="finding-decision-reasons">
-                              {detailDecisionReasons.map((reason) => (
-                                <div
-                                  data-tone={reason.tone}
-                                  key={`${reason.label}:${reason.detail}`}
-                                >
-                                  <dt>{reason.label}</dt>
-                                  <dd>{reason.detail}</dd>
-                                </div>
-                              ))}
-                            </dl>
-                            {detailReasonRows.length > 0 ? (
-                              <div className="finding-provider-reasons">
-                                <span>Provider explanation</span>
-                                <dl>
-                                  {detailReasonRows.map((reason) => (
-                                    <div
-                                      key={`${reason.label}:${reason.detail}`}
-                                    >
-                                      <dt>{labelize(reason.label)}</dt>
-                                      <dd>{reason.detail}</dd>
-                                    </div>
-                                  ))}
-                                </dl>
-                              </div>
-                            ) : null}
-                          </VpwSurfaceBody>
-                        </VpwSurface>
-
-                        <VpwSurface className="finding-decision-card finding-action-card">
-                          <VpwSurfaceHeader>
-                            <VpwSurfaceDescription>
-                              Remediation
-                            </VpwSurfaceDescription>
-                            <VpwSurfaceTitle>Decision plan</VpwSurfaceTitle>
-                          </VpwSurfaceHeader>
-                          <VpwSurfaceBody>
-                            <dl className="finding-decision-definition-list">
-                              <div>
-                                <dt>Recommended action</dt>
-                                <dd>
-                                  {findingRecommendedAction(
-                                    findingDetail,
-                                    findingExplanation,
-                                  )}
-                                </dd>
-                              </div>
-                              <div>
-                                <dt>SLA</dt>
-                                <dd>
-                                  {findingSlaLabel(findingDetail.priority)}
-                                </dd>
-                              </div>
-                              <div>
-                                <dt>Owner</dt>
-                                <dd>
-                                  {findingOwnerDetailLabel(
-                                    findingDetail,
-                                    detailOccurrences,
-                                  )}
-                                </dd>
-                              </div>
-                              <div>
-                                <dt>Next step</dt>
-                                <dd>{findingNextStepLabel(findingDetail)}</dd>
-                              </div>
-                              <div>
-                                <dt>Risk acceptance option</dt>
-                                <dd>
-                                  {detailWaiverEvidence
-                                    ? `${optionalText(detailWaiverEvidence.status)} — ${optionalText(detailWaiverEvidence.reason)}`
-                                    : "Available only with owner, expiry, approval reference, and compensating evidence."}
-                                </dd>
-                              </div>
-                            </dl>
-                            <div className="finding-decision-actions">
-                              <Button
-                                onClick={() =>
-                                  setFindingDetailReloadKey(
-                                    (value) => value + 1,
-                                  )
-                                }
-                                size="sm"
-                                type="button"
-                                variant="outline"
-                              >
-                                Refresh evidence
-                              </Button>
-                            </div>
-                          </VpwSurfaceBody>
-                        </VpwSurface>
-                      </section>
-
-                      <Tabs
-                        className="finding-detail-tabs-shell"
-                        value={findingDetailTab}
-                        onValueChange={(v) =>
-                          setFindingDetailTab(v as FindingDetailTab)
-                        }
-                      >
-                        <div className="finding-tabs-toolbar">
-                          <div className="finding-tabs-heading">
-                            <span>Decision evidence</span>
-                            <strong>
-                              Evidence, TTP context, and lifecycle
-                            </strong>
-                            <p>
-                              Provider-backed facts used to explain and defend
-                              the priority decision.
-                            </p>
-                          </div>
-                          <TabsList className="finding-detail-tabs-list">
-                            <TabsTrigger value="evidence">Evidence</TabsTrigger>
-                            <TabsTrigger value="ttp">TTP Context</TabsTrigger>
-                            <TabsTrigger value="history">History</TabsTrigger>
-                          </TabsList>
-                        </div>
-
-                        <TabsContent
-                          className="finding-detail-tab-panel"
-                          value="evidence"
-                        >
-                          <section
-                            className="finding-evidence-tab-layout"
-                            aria-label="Evidence workspace"
-                          >
-                            <section
-                              className="finding-evidence-grid"
-                              aria-label="Evidence summary"
-                            >
-                              {detailEvidenceRows.map((row, index) => (
-                                <VpwSurface
-                                  className={cn(
-                                    "finding-evidence-summary-card",
-                                    index === 0
-                                      ? "finding-evidence-summary-card-primary"
-                                      : undefined,
-                                  )}
-                                  key={row.label}
-                                >
-                                  <VpwSurfaceHeader>
-                                    <VpwSurfaceDescription>
-                                      {row.label}
-                                    </VpwSurfaceDescription>
-                                    <VpwSurfaceTitle>
-                                      {row.value}
-                                    </VpwSurfaceTitle>
-                                  </VpwSurfaceHeader>
-                                  <VpwSurfaceBody>
-                                    <p>{row.detail}</p>
-                                  </VpwSurfaceBody>
-                                </VpwSurface>
-                              ))}
-                            </section>
-
-                            <section
-                              className="finding-evidence-detail-grid"
-                              aria-label="Evidence detail"
-                            >
-                              <VpwSurface
-                                aria-label="Scanner occurrences"
-                                className="finding-tab-card finding-occurrences-card"
-                              >
-                                <VpwSurfaceHeader>
-                                  <div className="finding-card-heading">
-                                    <div>
-                                      <VpwSurfaceDescription>
-                                        Scanner evidence
-                                      </VpwSurfaceDescription>
-                                      <VpwSurfaceTitle>
-                                        Occurrences
-                                      </VpwSurfaceTitle>
-                                    </div>
-                                    <Badge variant="outline">
-                                      {detailOccurrences.length} source row(s)
-                                    </Badge>
-                                  </div>
-                                </VpwSurfaceHeader>
-                                <VpwSurfaceBody>
-                                  {detailOccurrences.length > 0 ? (
-                                    <div className="finding-detail-table-wrap">
-                                      <Table aria-label="Occurrences table">
-                                        <TableHeader>
-                                          <TableRow>
-                                            <TableHead>Source</TableHead>
-                                            <TableHead>Component</TableHead>
-                                            <TableHead>Asset / Owner</TableHead>
-                                            <TableHead>Evidence</TableHead>
-                                            <TableHead>Fix / VEX</TableHead>
-                                          </TableRow>
-                                        </TableHeader>
-                                        <TableBody>
-                                          {detailOccurrences.map(
-                                            (occurrence, index) => {
-                                              const fixVersions =
-                                                Array.isArray(
-                                                  occurrence.fix_versions,
-                                                ) &&
-                                                occurrence.fix_versions.length >
-                                                  0
-                                                  ? occurrence.fix_versions.join(
-                                                      ", ",
-                                                    )
-                                                  : stringValue(
-                                                      occurrence.fix_version,
-                                                    )
-                                              return (
-                                                <TableRow
-                                                  key={
-                                                    stringValue(
-                                                      occurrence.id,
-                                                    ) ??
-                                                    `occurrence-${index + 1}`
-                                                  }
-                                                >
-                                                  <TableCell>
-                                                    <span className="font-medium">
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.source_format,
-                                                        ) ??
-                                                          stringValue(
-                                                            occurrence.source,
-                                                          ),
-                                                      )}
-                                                    </span>
-                                                    <small>
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.source_record_id,
-                                                        ) ??
-                                                          stringValue(
-                                                            occurrence.raw_reference,
-                                                          ),
-                                                      )}
-                                                    </small>
-                                                  </TableCell>
-                                                  <TableCell>
-                                                    <span className="font-medium">
-                                                      {joinedValues([
-                                                        stringValue(
-                                                          occurrence.component_name,
-                                                        ),
-                                                        stringValue(
-                                                          occurrence.component_version,
-                                                        ),
-                                                      ])}
-                                                    </span>
-                                                    <small>
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.purl,
-                                                        ),
-                                                      )}
-                                                    </small>
-                                                  </TableCell>
-                                                  <TableCell>
-                                                    <span className="font-medium">
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.asset_ref,
-                                                        ) ??
-                                                          stringValue(
-                                                            occurrence.target_ref,
-                                                          ),
-                                                      )}
-                                                    </span>
-                                                    <small>
-                                                      {joinedValues([
-                                                        stringValue(
-                                                          occurrence.asset_owner,
-                                                        ),
-                                                        stringValue(
-                                                          occurrence.asset_business_service,
-                                                        ),
-                                                        labelize(
-                                                          stringValue(
-                                                            occurrence.asset_exposure,
-                                                          ),
-                                                        ),
-                                                      ])}
-                                                    </small>
-                                                  </TableCell>
-                                                  <TableCell>
-                                                    <span className="font-medium">
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.scanner,
-                                                        ),
-                                                      )}
-                                                    </span>
-                                                    <small>
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.raw_severity,
-                                                        ),
-                                                      )}
-                                                    </small>
-                                                  </TableCell>
-                                                  <TableCell>
-                                                    <span className="font-medium">
-                                                      {optionalText(
-                                                        fixVersions,
-                                                      )}
-                                                    </span>
-                                                    <small>
-                                                      {optionalText(
-                                                        stringValue(
-                                                          occurrence.vex_status,
-                                                        )
-                                                          ? labelize(
-                                                              stringValue(
-                                                                occurrence.vex_status,
-                                                              ),
-                                                            )
-                                                          : stringValue(
-                                                              occurrence.vex_justification,
-                                                            ),
-                                                      )}
-                                                    </small>
-                                                  </TableCell>
-                                                </TableRow>
-                                              )
-                                            },
-                                          )}
-                                        </TableBody>
-                                      </Table>
-                                    </div>
-                                  ) : (
-                                    <p className="text-sm text-muted-foreground">
-                                      No source occurrences have been recorded
-                                      for this finding.
-                                    </p>
-                                  )}
-                                </VpwSurfaceBody>
-                              </VpwSurface>
-
-                              <VpwSurface
-                                aria-label="Data quality notes"
-                                className="finding-tab-card finding-data-quality-card"
-                              >
-                                <VpwSurfaceHeader>
-                                  <VpwSurfaceDescription>
-                                    Provider data
-                                  </VpwSurfaceDescription>
-                                  <VpwSurfaceTitle>
-                                    Data quality notes
-                                  </VpwSurfaceTitle>
-                                </VpwSurfaceHeader>
-                                <VpwSurfaceBody>
-                                  {detailDataQualityRows.length > 0 ? (
-                                    <ul className="finding-data-quality-list">
-                                      {detailDataQualityRows.map((flag) => (
-                                        <li key={flag.key}>
-                                          <strong>{labelize(flag.code)}</strong>
-                                          <span>
-                                            {flag.source} /{" "}
-                                            {labelize(flag.severity)}
-                                          </span>
-                                          <p>{flag.message}</p>
-                                        </li>
-                                      ))}
-                                    </ul>
-                                  ) : (
-                                    <p className="text-sm text-muted-foreground">
-                                      No data quality flags recorded.
-                                    </p>
-                                  )}
-                                </VpwSurfaceBody>
-                              </VpwSurface>
-                            </section>
-                          </section>
-                        </TabsContent>
-
-                        <TabsContent
-                          className="finding-detail-tab-panel"
-                          value="ttp"
-                        >
-                          <VpwSurface
-                            aria-label="TTP Context"
-                            className="finding-tab-card finding-ttp-card"
-                          >
-                            <VpwSurfaceHeader>
-                              <div className="finding-card-heading">
-                                <div>
-                                  <VpwSurfaceDescription>
-                                    ATT&amp;CK
-                                  </VpwSurfaceDescription>
-                                  <VpwSurfaceTitle>TTP Context</VpwSurfaceTitle>
-                                </div>
-                                <Badge variant="outline">
-                                  {detailAttackSource
-                                    ?.toLowerCase()
-                                    .includes("demo")
-                                    ? "Curated demo mapping"
-                                    : detailAttackContext?.mapped
-                                      ? "Mapped context"
-                                      : "No approved mapping"}
-                                </Badge>
-                              </div>
-                            </VpwSurfaceHeader>
-                            <VpwSurfaceBody>
-                              {detailAttackEmpty ? (
-                                <section
-                                  aria-label="TTP context empty state"
-                                  className="finding-empty-panel"
-                                >
-                                  <div className="finding-empty-panel-heading">
-                                    <Badge variant="outline">
-                                      Defensive context
-                                    </Badge>
-                                    <strong>
-                                      No approved ATT&amp;CK mapping is stored
-                                      for this finding.
-                                    </strong>
-                                  </div>
-                                  <p>
-                                    Workbench does not infer tactics or
-                                    techniques for unmapped CVEs. Add a reviewed
-                                    CTID or curated local mapping before using
-                                    ATT&amp;CK context in queue decisions.
-                                  </p>
-                                </section>
-                              ) : (
-                                <>
-                                  <section
-                                    className="finding-ttp-context-hero finding-ttp-context-hero-simple"
-                                    aria-label="Threat informed context"
-                                  >
-                                    <div className="finding-ttp-main-copy">
-                                      <span>Threat informed context</span>
-                                      <h3>{detailAttackTechniqueLabel}</h3>
-                                      <p>
-                                        This mapping explains why the finding is
-                                        treated as an internet-facing Initial
-                                        Access risk. It supports remediation
-                                        priority and detection coverage review,
-                                        but does not prove exploitation.
-                                      </p>
-                                    </div>
-                                    <dl className="finding-ttp-main-facts">
-                                      <div>
-                                        <dt>Tactic</dt>
-                                        <dd>{detailAttackTacticLabel}</dd>
-                                      </div>
-                                      <div>
-                                        <dt>Confidence</dt>
-                                        <dd>
-                                          <Badge
-                                            className={cn(
-                                              detailAttackContext?.confidence ===
-                                                "high"
-                                                ? "bg-green-100 text-green-700 border-green-200"
-                                                : detailAttackContext?.confidence ===
-                                                    "low"
-                                                  ? "bg-red-100 text-red-700 border-red-200"
-                                                  : "bg-yellow-100 text-yellow-700 border-yellow-200",
-                                            )}
-                                          >
-                                            {attackConfidenceLabel(
-                                              detailAttackContext?.confidence,
-                                            )}
-                                          </Badge>
-                                        </dd>
-                                      </div>
-                                      <div>
-                                        <dt>Source</dt>
-                                        <dd>
-                                          {optionalText(detailAttackSource)}
-                                        </dd>
-                                      </div>
-                                      <div>
-                                        <dt>Coverage</dt>
-                                        <dd>{detailAttackCoverageStatus}</dd>
-                                      </div>
-                                    </dl>
-                                  </section>
-
-                                  <ol
-                                    className="finding-ttp-chain finding-ttp-chain-compact"
-                                    aria-label="Attack chain and decision flow"
-                                  >
-                                    <li data-tone="signal">
-                                      <span className="finding-ttp-chain-index">
-                                        1
-                                      </span>
-                                      <div>
-                                        <small>CVE signal</small>
-                                        <strong>High priority signal</strong>
-                                      </div>
-                                    </li>
-                                    <li data-tone="exposure">
-                                      <span className="finding-ttp-chain-index">
-                                        2
-                                      </span>
-                                      <div>
-                                        <small>Internet-facing asset</small>
-                                        <strong>Internet facing asset</strong>
-                                      </div>
-                                    </li>
-                                    <li data-tone="technique">
-                                      <span className="finding-ttp-chain-index">
-                                        3
-                                      </span>
-                                      <div>
-                                        <small>ATT&amp;CK technique</small>
-                                        <strong>
-                                          {detailAttackTechniqueId}
-                                        </strong>
-                                      </div>
-                                    </li>
-                                    <li data-tone="coverage">
-                                      <span className="finding-ttp-chain-index">
-                                        4
-                                      </span>
-                                      <div>
-                                        <small>Detection gap</small>
-                                        <strong>
-                                          {detailAttackCoverageStatus}
-                                        </strong>
-                                      </div>
-                                    </li>
-                                    <li data-tone="decision">
-                                      <span className="finding-ttp-chain-index">
-                                        5
-                                      </span>
-                                      <div>
-                                        <small>Remediation priority</small>
-                                        <strong>Emergency remediation</strong>
-                                      </div>
-                                    </li>
-                                  </ol>
-
-                                  <section
-                                    className="finding-ttp-decision-grid finding-ttp-decision-grid-simple"
-                                    aria-label="Threat informed decision context"
-                                  >
-                                    <article className="finding-ttp-narrative-card">
-                                      <span>Why this matters</span>
-                                      <strong>Decision support</strong>
-                                      <ul className="finding-ttp-meaning-list">
-                                        <li>
-                                          Frames the CVE as an Initial Access
-                                          risk.
-                                        </li>
-                                        <li>
-                                          Explains why internet exposure raises
-                                          urgency.
-                                        </li>
-                                        <li>
-                                          Connects remediation priority with
-                                          detection review.
-                                        </li>
-                                        <li>
-                                          Keeps the boundary clear: no proof of
-                                          exploitation.
-                                        </li>
-                                      </ul>
-                                    </article>
-                                    <article className="finding-ttp-narrative-card finding-ttp-actions-card">
-                                      <span>Recommended defensive actions</span>
-                                      <strong>
-                                        Close exposure and validate coverage
-                                      </strong>
-                                      <ul className="finding-ttp-checklist">
-                                        {(detailAttackActionItems.length > 0
-                                          ? detailAttackActionItems
-                                          : [
-                                              "Patch or mitigate the vulnerable service.",
-                                              "Restrict exposure while remediation is in progress.",
-                                              "Validate web, proxy, WAF, EDR and application telemetry.",
-                                              "Document detection coverage and residual risk.",
-                                            ]
-                                        ).map((item) => (
-                                          <li key={item}>{item}</li>
-                                        ))}
-                                      </ul>
-                                    </article>
-                                  </section>
-
-                                  <details className="finding-ttp-technical-evidence finding-ttp-technical-details">
-                                    <summary className="finding-ttp-technical-heading">
-                                      <div>
-                                        <span>Secondary evidence</span>
-                                        <strong>
-                                          Technical mapping details
-                                        </strong>
-                                      </div>
-                                      <Badge variant="outline">
-                                        Source, confidence, rationale
-                                      </Badge>
-                                    </summary>
-
-                                    <div className="finding-ttp-technical-body">
-                                      <p>
-                                        {optionalText(detailAttackRationale)}
-                                      </p>
-                                      <div className="finding-detail-table-wrap">
-                                        <Table aria-label="TTP Context techniques">
-                                          <TableHeader>
-                                            <TableRow>
-                                              <TableHead>Technique</TableHead>
-                                              <TableHead>Tactic</TableHead>
-                                              <TableHead>Confidence</TableHead>
-                                              <TableHead>Source</TableHead>
-                                              <TableHead>Rationale</TableHead>
-                                              <TableHead>
-                                                Defensive actions
-                                              </TableHead>
-                                            </TableRow>
-                                          </TableHeader>
-                                          <TableBody>
-                                            {detailAttackTechniques.map(
-                                              (technique) => {
-                                                const actionItems =
-                                                  defensiveActionItems(
-                                                    technique.defensive_note,
-                                                  )
-                                                return (
-                                                  <TableRow
-                                                    key={technique.technique_id}
-                                                  >
-                                                    <TableCell>
-                                                      <span className="font-medium">
-                                                        {technique.technique_id}
-                                                      </span>
-                                                      <small>
-                                                        {optionalText(
-                                                          technique.name,
-                                                        )}
-                                                      </small>
-                                                    </TableCell>
-                                                    <TableCell>
-                                                      {attackTacticsLabel(
-                                                        technique.tactics,
-                                                      )}
-                                                    </TableCell>
-                                                    <TableCell>
-                                                      <Badge
-                                                        className={cn(
-                                                          technique.confidence ===
-                                                            "high"
-                                                            ? "bg-green-100 text-green-700 border-green-200"
-                                                            : technique.confidence ===
-                                                                "low"
-                                                              ? "bg-red-100 text-red-700 border-red-200"
-                                                              : "bg-yellow-100 text-yellow-700 border-yellow-200",
-                                                        )}
-                                                      >
-                                                        {attackConfidenceLabel(
-                                                          technique.confidence,
-                                                        )}
-                                                      </Badge>
-                                                    </TableCell>
-                                                    <TableCell>
-                                                      {optionalText(
-                                                        attackSourceLabel(
-                                                          technique.source,
-                                                          detailAttackContext,
-                                                        ),
-                                                      )}
-                                                    </TableCell>
-                                                    <TableCell>
-                                                      {optionalText(
-                                                        technique.rationale,
-                                                      )}
-                                                    </TableCell>
-                                                    <TableCell>
-                                                      {actionItems.length >
-                                                      1 ? (
-                                                        <ul className="finding-ttp-actions-list">
-                                                          {actionItems.map(
-                                                            (item) => (
-                                                              <li key={item}>
-                                                                {item}
-                                                              </li>
-                                                            ),
-                                                          )}
-                                                        </ul>
-                                                      ) : (
-                                                        optionalText(
-                                                          technique.defensive_note,
-                                                        )
-                                                      )}
-                                                    </TableCell>
-                                                  </TableRow>
-                                                )
-                                              },
-                                            )}
-                                          </TableBody>
-                                        </Table>
-                                      </div>
-                                    </div>
-                                  </details>
-                                </>
-                              )}
-                              {detailAttackEmpty ? (
-                                <section
-                                  className="finding-detection-note"
-                                  aria-label="Detection coverage"
-                                >
-                                  <span>Detection coverage</span>
-                                  <p>
-                                    {optionalText(
-                                      detailAttackContext?.defensive_note ??
-                                        "Coverage controls are not connected to this finding yet. Record detection and mitigation evidence when available.",
-                                    )}
-                                  </p>
-                                </section>
-                              ) : null}
-                              <p className="finding-defensive-note">
-                                Defensive context only. No exploit steps,
-                                payloads, PoC guidance, active probing, or
-                                offensive procedure instructions.
-                              </p>
-                            </VpwSurfaceBody>
-                          </VpwSurface>
-                        </TabsContent>
-
-                        <TabsContent
-                          className="finding-detail-tab-panel"
-                          value="history"
-                        >
-                          <section
-                            className="finding-history-timeline"
-                            aria-label="Finding history"
-                          >
-                            {detailHistoryRows.map((row, index) => (
-                              <VpwSurface
-                                className="finding-tab-card finding-history-event"
-                                key={row.label}
-                              >
-                                <VpwSurfaceHeader>
-                                  <span className="finding-history-step">
-                                    {index + 1}
-                                  </span>
-                                  <div>
-                                    <VpwSurfaceDescription>
-                                      {row.label}
-                                    </VpwSurfaceDescription>
-                                    <VpwSurfaceTitle>
-                                      {row.value}
-                                    </VpwSurfaceTitle>
-                                  </div>
-                                </VpwSurfaceHeader>
-                                <VpwSurfaceBody>
-                                  <p>{row.detail}</p>
-                                </VpwSurfaceBody>
-                              </VpwSurface>
-                            ))}
-                          </section>
-                          {detailWaiverEvidence ? (
-                            <VpwSurface
-                              aria-label="Accepted risk"
-                              className="finding-tab-card finding-accepted-risk-card"
-                            >
-                              <VpwSurfaceHeader>
-                                <VpwSurfaceDescription>
-                                  Risk acceptance
-                                </VpwSurfaceDescription>
-                                <VpwSurfaceTitle>Accepted risk</VpwSurfaceTitle>
-                              </VpwSurfaceHeader>
-                              <VpwSurfaceBody>
-                                <dl className="finding-decision-definition-list compact">
-                                  <div>
-                                    <dt>Owner</dt>
-                                    <dd>
-                                      {optionalText(detailWaiverEvidence.owner)}
-                                    </dd>
-                                  </div>
-                                  <div>
-                                    <dt>Reason</dt>
-                                    <dd>
-                                      {optionalText(
-                                        detailWaiverEvidence.reason,
-                                      )}
-                                    </dd>
-                                  </div>
-                                  <div>
-                                    <dt>Expires</dt>
-                                    <dd>
-                                      {optionalText(
-                                        detailWaiverEvidence.expiresOn,
-                                      )}
-                                    </dd>
-                                  </div>
-                                  <div>
-                                    <dt>Review</dt>
-                                    <dd>
-                                      {optionalText(
-                                        detailWaiverEvidence.reviewOn,
-                                      )}
-                                    </dd>
-                                  </div>
-                                  <div>
-                                    <dt>Scope</dt>
-                                    <dd>
-                                      {optionalText(
-                                        detailWaiverEvidence.matchedScope ??
-                                          detailWaiverEvidence.scope,
-                                      )}
-                                    </dd>
-                                  </div>
-                                  <div>
-                                    <dt>Approval</dt>
-                                    <dd>
-                                      {optionalText(
-                                        detailWaiverEvidence.approvalRef,
-                                      )}
-                                    </dd>
-                                  </div>
-                                </dl>
-                              </VpwSurfaceBody>
-                            </VpwSurface>
-                          ) : null}
-                        </TabsContent>
-                      </Tabs>
-                    </>
-                  ) : null}
-                </section>
+                <FindingDetailRoute
+                  error={findingDetailError}
+                  explanation={findingExplanation}
+                  explanationWarning={findingExplanationWarning}
+                  finding={findingDetail}
+                  loading={findingDetailLoading}
+                  onRefresh={refreshFindingDetail}
+                  onTabChange={setFindingDetailTab}
+                  tab={findingDetailTab}
+                />
               ) : currentPath === "/providers" ? (
                 <ProvidersRouteContainer
                   onRefreshProviderStatus={refreshProviderStatus}
